BUG: revert b0dcdda71f96: Vector data associated value CAN be unsigned short int

+++ Examples/IO/VectorDataIOExample.cxx
@@ -69,7 +69,7 @@ int main(int argc, char * argv[])
     return EXIT_FAILURE;
   }
 
-  typedef double PixelType;
+  typedef unsigned short int PixelType;
 
 //  Software Guide : BeginLatex
   //
